@@ -1,15 +1,15 @@
 <Project Sdk="Microsoft.NET.Sdk.Web">
   <PropertyGroup>
-    <TargetFramework>net8.0</TargetFramework>
+    <TargetFramework>net9.0</TargetFramework>
     <Nullable>enable</Nullable>
     <ImplicitUsings>enable</ImplicitUsings>
     <Version>2.1.0</Version>
-    <LangVersion>12</LangVersion>
+    <LangVersion>13</LangVersion>
   </PropertyGroup>
   <ItemGroup>
     <ProjectReference Include="..\DotnetActuatorMiddleware\DotnetActuatorMiddleware.csproj"/>
   </ItemGroup>
   <ItemGroup>
-    <PackageReference Include="Quartz" Version="3.8.1"/>
+    <PackageReference Include="Quartz" Version="3.15.0" />
   </ItemGroup>
 </Project>
