@@ -1,12 +1,12 @@
 <Project Sdk="Microsoft.NET.Sdk">
   <PropertyGroup>
-    <TargetFramework>net8.0</TargetFramework>
+    <TargetFramework>net9.0</TargetFramework>
     <ImplicitUsings>enable</ImplicitUsings>
     <Nullable>enable</Nullable>
     <Version>2.1.0</Version>
     <PackageReadmeFile>README.md</PackageReadmeFile>
     <Description>Provides actuator style health and information endpoints for .NET web applications.</Description>
-    <LangVersion>12</LangVersion>
+    <LangVersion>13</LangVersion>
   </PropertyGroup>
   <ItemGroup>
     <AssemblyAttribute Include="System.Runtime.CompilerServices.InternalsVisibleToAttribute">
@@ -21,14 +21,14 @@
     <None Include="../../README.md" Pack="true" PackagePath="\"/>
   </ItemGroup>
   <ItemGroup>
-    <PackageReference Include="Elastic.Clients.Elasticsearch" Version="8.12.0"/>
-    <PackageReference Include="IPAddressRange" Version="6.0.0"/>
-    <PackageReference Include="Newtonsoft.Json" Version="13.0.3"/>
-    <PackageReference Include="MySql.Data" Version="8.3.0">
+    <PackageReference Include="Elastic.Clients.Elasticsearch" Version="9.1.5" />
+    <PackageReference Include="IPAddressRange" Version="6.2.0" />
+    <PackageReference Include="MySql.Data" Version="9.4.0">
       <PrivateAssets>all</PrivateAssets>
     </PackageReference>
-    <PackageReference Include="Quartz" Version="3.8.1">
+    <PackageReference Include="Quartz" Version="3.15.0">
       <PrivateAssets>all</PrivateAssets>
     </PackageReference>
+    <PackageReference Include="System.Text.Json" Version="9.0.8" />
   </ItemGroup>
 </Project>
