@@ -1,10 +1,8 @@
+using System.Text.Json;
+
 using DotnetActuatorMiddleware.Endpoints;
-using DotnetActuatorMiddleware.Health;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Http;
-
-using Newtonsoft.Json;
-using Newtonsoft.Json.Linq;
 
 namespace DotnetActuatorMiddleware;
 
@@ -16,19 +14,19 @@ public static class ActuatorMiddleware
         {
             if (context.Request.Path.Value is not null && context.Request.Path.Value == "/health")
             {
-                context.Response.Headers["Content-Type"] = "application/json";
+                context.Response.Headers.ContentType = "application/json";
 
                 var healthEndpoint = new HealthEndpoint(ipAllowListEnabled);
                 
                 // Check if request is coming from an allowed IP
-                if (healthEndpoint.IpAllowListEnabled && context.Connection.RemoteIpAddress is not null && !healthEndpoint.IpIsAllowed(context.Connection.RemoteIpAddress))
+                if (healthEndpoint.IpAllowListEnabled && context.Connection.RemoteIpAddress is not null && !ActuatorEndpoint.IpIsAllowed(context.Connection.RemoteIpAddress))
                 {
                     context.Response.StatusCode = 401;
-                    await context.Response.WriteAsync(JsonConvert.SerializeObject(new { message = "Forbidden" }));
+                    await context.Response.WriteAsync(JsonSerializer.Serialize(new { message = "Forbidden" }));
                     return;
                 }
 
-                var applicationHealthStatus = await Task.Run(() => healthEndpoint.GetHealth());
+                var applicationHealthStatus = await Task.Run(() => HealthEndpoint.GetHealth());
 
                 if (!applicationHealthStatus.IsHealthy)
                 {
@@ -36,15 +34,24 @@ public static class ActuatorMiddleware
                     context.Response.StatusCode = 503;
                 }
                 
-                // Assemble JSON payload
-                var jsonResponseRoot = new JObject();
-                jsonResponseRoot.Add("healthy", applicationHealthStatus.IsHealthy);
-                foreach ((string? key, HealthResponse? value) in applicationHealthStatus.Results)
+                // System.text.json doesn't appear to flatten the same way that Newtonsoft did so we do that here
+                // while trying to be somewhat memory efficient about it
+                await using var json = new Utf8JsonWriter(
+                    context.Response.BodyWriter,
+                    new JsonWriterOptions { SkipValidation = true }
+                );
+                
+                json.WriteStartObject();
+                json.WriteBoolean("healthy", applicationHealthStatus.IsHealthy);
+                
+                foreach (var kv in applicationHealthStatus.Results)
                 {
-                    jsonResponseRoot.Add(key, JToken.FromObject(value));
+                    json.WritePropertyName(kv.Key);
+                    JsonSerializer.Serialize(json, kv.Value);
                 }
 
-                await context.Response.WriteAsync(JsonConvert.SerializeObject(jsonResponseRoot));
+                json.WriteEndObject();
+                await json.FlushAsync();
             }
             else
             {
@@ -59,21 +66,21 @@ public static class ActuatorMiddleware
         {
             if (context.Request.Path.Value is not null && context.Request.Path.Value == "/info")
             {
-                context.Response.Headers["Content-Type"] = "application/json";
+                context.Response.Headers.ContentType = "application/json";
 
                 var infoEndpoint = new InfoEndpoint(ipAllowListEnabled);
                 
                 // Check if request is coming from an allowed IP
-                if (infoEndpoint.IpAllowListEnabled && context.Connection.RemoteIpAddress is not null && !infoEndpoint.IpIsAllowed(context.Connection.RemoteIpAddress))
+                if (infoEndpoint.IpAllowListEnabled && context.Connection.RemoteIpAddress is not null && !ActuatorEndpoint.IpIsAllowed(context.Connection.RemoteIpAddress))
                 {
                     context.Response.StatusCode = 401;
-                    await context.Response.WriteAsync(JsonConvert.SerializeObject(new { message = "Forbidden" }));
+                    await context.Response.WriteAsync(JsonSerializer.Serialize(new { message = "Forbidden" }));
                     return;
                 }
 
-                var infoReponse = infoEndpoint.GetInfo();
+                var infoReponse = InfoEndpoint.GetInfo();
 
-                await context.Response.WriteAsync(JsonConvert.SerializeObject(infoReponse));
+                await context.Response.WriteAsync(JsonSerializer.Serialize(infoReponse));
             }
             else
             {
@@ -88,19 +95,19 @@ public static class ActuatorMiddleware
         {
             if (context.Request.Path.Value is not null && context.Request.Path.Value == "/env")
             {
-                context.Response.Headers["Content-Type"] = "application/json";
+                context.Response.Headers.ContentType = "application/json";
 
                 var environmentEndpoint = new EnvironmentEndpoint(ipAllowListEnabled);
                 
                 // Check if request is coming from an allowed IP
-                if (environmentEndpoint.IpAllowListEnabled && context.Connection.RemoteIpAddress is not null && !environmentEndpoint.IpIsAllowed(context.Connection.RemoteIpAddress))
+                if (environmentEndpoint.IpAllowListEnabled && context.Connection.RemoteIpAddress is not null && !ActuatorEndpoint.IpIsAllowed(context.Connection.RemoteIpAddress))
                 {
                     context.Response.StatusCode = 401;
-                    await context.Response.WriteAsync(JsonConvert.SerializeObject(new { message = "Forbidden" }));
+                    await context.Response.WriteAsync(JsonSerializer.Serialize(new { message = "Forbidden" }));
                     return;
                 }
 
-                await context.Response.WriteAsync(JsonConvert.SerializeObject(environmentEndpoint.GetEnvironment(), new JsonSerializerSettings { StringEscapeHandling = StringEscapeHandling.EscapeNonAscii }));
+                await context.Response.WriteAsync(JsonSerializer.Serialize(EnvironmentEndpoint.GetEnvironment()));
 
             }
             else
@@ -116,19 +123,19 @@ public static class ActuatorMiddleware
         {
             if (context.Request.Path.Value is not null && context.Request.Path.Value == "/quartz")
             {
-                context.Response.Headers["Content-Type"] = "application/json";
+                context.Response.Headers.ContentType = "application/json";
 
                 var quartzEndpoint = new QuartzEndpoint(ipAllowListEnabled);
                 
                 // Check if request is coming from an allowed IP
-                if (quartzEndpoint.IpAllowListEnabled && context.Connection.RemoteIpAddress is not null && !quartzEndpoint.IpIsAllowed(context.Connection.RemoteIpAddress))
+                if (quartzEndpoint.IpAllowListEnabled && context.Connection.RemoteIpAddress is not null && !ActuatorEndpoint.IpIsAllowed(context.Connection.RemoteIpAddress))
                 {
                     context.Response.StatusCode = 401;
-                    await context.Response.WriteAsync(JsonConvert.SerializeObject(new { message = "Forbidden" }));
+                    await context.Response.WriteAsync(JsonSerializer.Serialize(new { message = "Forbidden" }));
                     return;
                 }
 
-                await context.Response.WriteAsync(JsonConvert.SerializeObject(quartzEndpoint.GetSchedulerStatus()));
+                await context.Response.WriteAsync(JsonSerializer.Serialize(quartzEndpoint.GetSchedulerStatus()));
 
             }
             else
