@@ -2,38 +2,38 @@ namespace DotnetActuatorMiddleware.Endpoints;
 
 public class QuartzEndpointResponse
 {
-    public Dictionary<string, QuartzEndpointScheduler> Schedulers = new Dictionary<string, QuartzEndpointScheduler>();
+    public Dictionary<string, QuartzEndpointScheduler> Schedulers { get; set; } = new Dictionary<string, QuartzEndpointScheduler>();
 }
 
 public class QuartzEndpointScheduler
 {
-    public string? SchedulerStatus;
-    public List<QuartzEndpointJob> Jobs = new List<QuartzEndpointJob>();
+    public string? SchedulerStatus { get; set; }
+    public List<QuartzEndpointJob> Jobs { get; set; } = [];
 }
 
 public class QuartzEndpointJob
 {
-    public string? Name;
-    public string? Group;
-    public string? Description;
-    public string? JobClass;
-    public bool? LastRunSuccessful = null;
-    public DateTimeOffset? LastErrorTimeUtc;
-    public string? LastRunErrorMessage = null;
-    public object? LastRunOutput = null;
-    public bool ConcurrentExecutionAllowed;
-    public bool PersistJobData;
-    public List<QuartzEndpointJobTriggers> Triggers = new List<QuartzEndpointJobTriggers>();
+    public string? Name { get; set; }
+    public string? Group { get; set; }
+    public string? Description { get; set; }
+    public string? JobClass { get; set; }
+    public bool? LastRunSuccessful { get; set; }
+    public DateTimeOffset? LastErrorTimeUtc { get; set; }
+    public string? LastRunErrorMessage { get; set; }
+    public object? LastRunOutput { get; set; }
+    public bool ConcurrentExecutionAllowed { get; set; }
+    public bool PersistJobData { get; set; }
+    public List<QuartzEndpointJobTriggers> Triggers  { get; set; } = [];
 }
 
 public struct QuartzEndpointJobTriggers
 {
-    public string Name;
-    public string Group;
-    public string? Description;
-    public DateTimeOffset? LastFireTimeUtc;
-    public DateTimeOffset? NextFireTimeUtc;
-    public DateTimeOffset? FinalFireTimeUtc;
-    public DateTimeOffset StartTimeUtc;
-    public DateTimeOffset? EndTimeUtc;
+    public string Name { get; set; }
+    public string Group { get; set; }
+    public string? Description { get; set; }
+    public DateTimeOffset? LastFireTimeUtc { get; set; }
+    public DateTimeOffset? NextFireTimeUtc { get; set; }
+    public DateTimeOffset? FinalFireTimeUtc { get; set; }
+    public DateTimeOffset StartTimeUtc { get; set; }
+    public DateTimeOffset? EndTimeUtc { get; set; }
 }
