@@ -2,6 +2,6 @@ namespace DotnetActuatorMiddleware.Endpoints;
 
 public struct InfoEndpointResponse
 {
-    public string? Name;
-    public string? Version;
+    public string? Name { get; set; }
+    public string? Version { get; set; }
 }
