@@ -1,8 +1,6 @@
 using Elastic.Clients.Elasticsearch;
 using Elastic.Transport;
 
-using Newtonsoft.Json;
-
 namespace DotnetActuatorMiddleware.Health.Checks;
 
 public static class ElasticsearchHealthCheck
@@ -14,7 +12,7 @@ public static class ElasticsearchHealthCheck
     /// <param name="timeoutSecs">Elasticsearch request timeout in seconds</param>
     /// <param name="username">Elasticsearch username to authenticate with</param>
     /// <param name="password">Elasticsearch password to authenticate with</param>
-    /// <param name="serverCertificateValidation">Whether or not to validate the SSL certificate returned by the server</param>
+    /// <param name="serverCertificateValidation">Whether to validate the SSL certificate returned by the server</param>
     /// <returns>A <see cref="HealthResponse"/> object that contains the return status of this health check</returns>
     public static HealthResponse CheckHealth(Uri[] servers, int timeoutSecs = 5, string? username = null, string? password = null, bool serverCertificateValidation = true)
     {
@@ -26,14 +24,12 @@ public static class ElasticsearchHealthCheck
                 .RequestTimeout(TimeSpan.FromSeconds(timeoutSecs))
                 .PingTimeout(TimeSpan.FromSeconds(timeoutSecs));
 
-            if (!serverCertificateValidation)
-            {
+            if (!serverCertificateValidation) {
                 // If server certificate validation is disabled then always return true when calling the validation callback
                 clientSettings.ServerCertificateValidationCallback((o, certificate, arg3, arg4) => true);
             }
 
-            if (!String.IsNullOrWhiteSpace(username) && !String.IsNullOrWhiteSpace(password))
-            {
+            if (!String.IsNullOrWhiteSpace(username) && !String.IsNullOrWhiteSpace(password)) {
                 clientSettings.Authentication(new BasicAuthentication(username, password));
             }
             
@@ -41,18 +37,15 @@ public static class ElasticsearchHealthCheck
             var lowlevelClient = new ElasticsearchClient(clientSettings);
 
             var clusterHealthResponse = lowlevelClient.Cluster.Health();
+            var healthResponse = new ElasticsearchHealthCheckResponse {
+                ClusterName = clusterHealthResponse.ClusterName, Status = clusterHealthResponse.Status.ToString(),
+            };
 
-            if (!clusterHealthResponse.ApiCallDetails.HasSuccessfulStatusCode)
-            {
-                return HealthResponse.Unhealthy(clusterHealthResponse.ApiCallDetails.OriginalException);
+            if (!clusterHealthResponse.ApiCallDetails.HasSuccessfulStatusCode) {
+                return clusterHealthResponse.ApiCallDetails.OriginalException is null ? HealthResponse.Unhealthy() : HealthResponse.Unhealthy(clusterHealthResponse.ApiCallDetails.OriginalException);
             }
 
-            if (clusterHealthResponse.Status != HealthStatus.Green)
-            {
-                return HealthResponse.Unhealthy(new ElasticsearchHealthCheckResponse { ClusterName = clusterHealthResponse.ClusterName, Status = clusterHealthResponse.Status.ToString() });
-            }
-            
-            return HealthResponse.Healthy(new ElasticsearchHealthCheckResponse { ClusterName = clusterHealthResponse.ClusterName, Status = clusterHealthResponse.Status.ToString() });
+            return clusterHealthResponse.Status != HealthStatus.Green ? HealthResponse.Unhealthy(healthResponse) : HealthResponse.Healthy(healthResponse);
         }
         catch (Exception e)
         {
