@@ -7,7 +7,7 @@ namespace DotnetActuatorMiddleware.Util;
 ///
 /// This class also allows application configuration to be exposed via the environment endpoint.
 /// </summary>
-public class ConfigurationRegistry
+public static class ConfigurationRegistry
 {
     internal static Dictionary<string, IConfigurationRoot> Sources { get; } = new Dictionary<string, IConfigurationRoot>();
     
@@ -18,12 +18,10 @@ public class ConfigurationRegistry
     /// </summary>
     public static void AddConfigurationSource(IConfigurationRoot configuration, string sourceName)
     {
-        if (Sources.ContainsKey(sourceName))
+        if (!Sources.TryAdd(sourceName, configuration))
         {
             throw new InvalidOperationException($"There is already a configuration source registered with the name {sourceName}");
         }
-
-        Sources.Add(sourceName, configuration);
     }
     
     /// <summary>
@@ -32,12 +30,10 @@ public class ConfigurationRegistry
     /// </summary>
     public static void RemoveConfigurationSource(string sourceName)
     {
-        if (!Sources.ContainsKey(sourceName))
+        if (!Sources.Remove(sourceName))
         {
             throw new InvalidOperationException($"There is no configuration source registered with the name {sourceName}");
         }
-
-        Sources.Remove(sourceName);
     }
     
     /// <summary>
@@ -57,12 +53,7 @@ public class ConfigurationRegistry
     /// <returns>An <see cref="IConfigurationSection"/> instance</returns>
     public static IConfigurationSection GetConfigurationSection(string sourceName, string key)
     {
-        if (!Sources.ContainsKey(sourceName))
-        {
-            throw new InvalidOperationException($"There is no configuration source registered with the name {sourceName}");
-        }
-
-        return Sources[sourceName].GetSection(key);
+        return !Sources.TryGetValue(sourceName, out IConfigurationRoot? value) ? throw new InvalidOperationException($"There is no configuration source registered with the name {sourceName}") : value.GetSection(key);
     }
     
     /// <summary>
@@ -72,12 +63,8 @@ public class ConfigurationRegistry
     /// </summary>
     public static string GetKey(string sourceName, string key)
     {
-        if (!Sources.ContainsKey(sourceName))
-        {
-            throw new InvalidOperationException($"There is no configuration source registered with the name {sourceName}");
-        }
-        
-        return Sources[sourceName].GetSection(key).Value;
+        // GetSection can never return null according to API docs, it'll just return an empty IConfigurationSection
+        return (!Sources.TryGetValue(sourceName, out IConfigurationRoot? value) ? throw new InvalidOperationException($"There is no configuration source registered with the name {sourceName}") : value.GetSection(key).Value)!;
     }
 
     /// <summary>
@@ -88,12 +75,12 @@ public class ConfigurationRegistry
     /// </summary>
     public static void SetKey(string sourceName, string key, string value)
     {
-        if (!Sources.ContainsKey(sourceName))
+        if (!Sources.TryGetValue(sourceName, out IConfigurationRoot? source))
         {
             throw new InvalidOperationException($"There is no configuration source registered with the name {sourceName}");
         }
-        
-        Sources[sourceName].GetSection(key).Value = value;
+
+        source.GetSection(key).Value = value;
     }
     
     /// <summary>
@@ -103,13 +90,13 @@ public class ConfigurationRegistry
     /// </summary>
     public static Dictionary<string, string> GetAllValuesFromSource(string sourceName)
     {
-        if (!Sources.ContainsKey(sourceName))
+        if (!Sources.TryGetValue(sourceName, out IConfigurationRoot? value))
         {
             throw new InvalidOperationException("No configuration source registered with name "+ sourceName);
         }
             
         // Get children of this source and return them
-        IConfiguration sourceConfiguration = Sources[sourceName];
+        IConfiguration sourceConfiguration = value;
         Dictionary<string, string> valuesDictionary = RecurseConfig(sourceConfiguration);
 
         //Dictionary<string, string> valuesDictionary = sourceConfiguration.GetChildren().ToDictionary(child => child.Key, child => child.Value);
@@ -122,17 +109,17 @@ public class ConfigurationRegistry
 
         foreach (var child in source.GetChildren())
         {
-            if (child.GetChildren().Count() != 0)
+            if (child.GetChildren().Any())
             {
                 result = result.Concat(RecurseConfig(child)).GroupBy(d => d.Key).ToDictionary(d => d.Key, d => d.First().Value);
             }
 
-            if (child.GetChildren().Count() != 0 && string.IsNullOrEmpty(child.Value))
+            if (child.GetChildren().Any() && string.IsNullOrEmpty(child.Value))
             {
                 continue;
             }
 
-            result.Add(child.Path, child.Value);
+            result.Add(child.Path, child.Value!);
 
         }
 
